refactor: better use of build tags for minimal binary
