+++ pkg/extensions/events/common.go
@@ -1,43 +1,30 @@
 package events
 
 import (
-	"crypto/tls"
-	"crypto/x509"
-	"os"
 	"time"
-
-	eventsconf "zotregistry.dev/zot/pkg/extensions/config/events"
 )
 
 const (
 	DefaultHTTPTimeout = 30 * time.Second
+	EventSource        = "zotregistry.dev"
 )
 
-func getTLSConfig(config eventsconf.SinkConfig) (*tls.Config, error) {
-	tlsConfig := &tls.Config{
-		MinVersion: tls.VersionTLS12,
-	}
+type EventType string
 
-	if config.TLSConfig.CACertFile != "" {
-		caCert, err := os.ReadFile(config.TLSConfig.CACertFile)
-		if err != nil {
-			return nil, err
-		}
+const (
+	ImageUpdatedEventType      EventType = "zotregistry.image.updated"
+	ImageDeletedEventType      EventType = "zotregistry.image.deleted"
+	ImageLintFailedEventType   EventType = "zotregistry.image.lint_failed"
+	RepositoryCreatedEventType EventType = "zotregistry.repository.created"
+)
 
-		caCertPool := x509.NewCertPool()
-		if !caCertPool.AppendCertsFromPEM(caCert) {
-			return nil, err
-		}
-		tlsConfig.RootCAs = caCertPool
-	}
+func (e EventType) String() string {
+	return string(e)
+}
 
-	if config.TLSConfig.CertFile != "" && config.TLSConfig.KeyFile != "" {
-		cert, err := tls.LoadX509KeyPair(config.TLSConfig.CertFile, config.TLSConfig.KeyFile)
-		if err != nil {
-			return nil, err
-		}
-		tlsConfig.Certificates = []tls.Certificate{cert}
-	}
-
-	return tlsConfig, nil
+type Recorder interface {
+	RepositoryCreated(name string) error
+	ImageUpdated(name, reference, digest, mediaType, manifest string) error
+	ImageDeleted(name, reference, digest, mediaType string) error
+	ImageLintFailed(name, reference, digest, mediaType, manifest string) error
 }
