+++ pkg/extensions/events/extension.go
@@ -4,9 +4,16 @@
 package events
 
 import (
+	cloudevents "github.com/cloudevents/sdk-go/v2"
+
 	zerr "zotregistry.dev/zot/errors"
 	"zotregistry.dev/zot/pkg/log"
 )
+
+type Sink interface {
+	Emit(*cloudevents.Event) cloudevents.Result
+	Close() error
+}
 
 func NewRecorder(logger log.Logger, sinks ...Sink) (Recorder, error) {
 	if sinks == nil {
